remaining cycle-stepped 6502 fixes
