+++ examples/sokol/CMakeLists.txt
@@ -53,14 +53,6 @@ fips_begin_app(atom-ui windowed)
     endif()
     fips_deps(roms common ui)
 fips_end_app()
-fips_begin_app(atomx windowed)
-    fips_vs_warning_level(3)
-    fips_files(atomx.c)
-    if (FIPS_IOS)
-        fips_files(ios-info.plist)
-    endif()
-    fips_deps(roms common)
-fips_end_app()
 target_compile_definitions(atom-ui PRIVATE CHIPS_USE_UI)
 
 fips_ide_group(examples/c64)
