+++ tests/CMakeLists.txt
@@ -50,24 +50,9 @@ fips_begin_app(m6502-nestest cmdline)
     fips_generate(FROM dump.yml TYPE dump HEADER dump.h)
 fips_end_app()
 
-fips_begin_app(m6502x-nestest cmdline)
-    fips_vs_warning_level(3)
-    fips_files(m6502x-nestest.c)
-    fips_dir(nestest)
-    fips_generate(FROM nestest.log.txt TYPE nestestlog HEADER nestestlog.h)
-    fips_generate(FROM dump.yml TYPE dump HEADER dump.h)
-fips_end_app()
-
 fips_begin_app(m6502-wltest cmdline)
     fips_vs_warning_level(3)
     fips_files(m6502-wltest.c)
-    fips_dir(testsuite-2.15/bin)
-    fips_generate(FROM dump.yml TYPE dump HEADER dump.h)
-fips_end_app()
-
-fips_begin_app(m6502x-wltest cmdline)
-    fips_vs_warning_level(3)
-    fips_files(m6502x-wltest.c)
     fips_dir(testsuite-2.15/bin)
     fips_generate(FROM dump.yml TYPE dump HEADER dump.h)
 fips_end_app()
